Remove HyperCoreServiceConfigProvider and rename to HyperCoreConfig
- Removed HyperCoreServiceConfigProvider.swift
- Renamed HyperCoreServiceConfig to HyperCoreConfig
- Added static create() method as extension to HyperCoreConfig
- Updated all references throughout the codebase
